Refactor ChainSignature and NativeNEAR components for improved routing and UI
- Updated ChainSignatureDashboard to use the correct path for supported chains data.
- Changed navigation routes in ChainSignatureDashboard and NativeNearDashboard to include mode parameters for better context.
- Removed the obsolete ChainSend page and integrated its functionality into the main send page.
- Simplified the send page to conditionally render either ChainSignatureSend or NativeNearSend based on the selected mode.
- Enhanced user experience by providing clear feedback on the transaction process and improving the overall layout.
