--- pages/chain-send.js
@@ -1,5 +1,0 @@
-import ChainSignatureSend from '../components/ChainSignature/ChainSignatureSend';
-
-export default function ChainSendPage() {
-  return <ChainSignatureSend />;
-}
